requirements.txt
pillow>=11.1.0,<11.2
PyMuPDF>=1.25.5,<1.26
